Added AbstractOracle
